cleaning up tests adding get active project api
