+++ api/specs/webserver/v0/openapi.yaml
@@ -117,6 +117,9 @@ paths:
   /projects:
     $ref: './openapi-projects.yaml#/paths/~1projects'
 
+  /projects/active:
+    $ref: './openapi-projects.yaml#/paths/~1projects~1active'
+
   /projects/{project_id}:
     $ref: './openapi-projects.yaml#/paths/~1projects~1{project_id}'
 
@@ -124,10 +127,7 @@ paths:
     $ref: './openapi-projects.yaml#/paths/~1projects~1{project_id}~1open'
 
   /projects/{project_id}:close:
-    $ref: './openapi-projects.yaml#/paths/~1projects~1{project_id}~1close'
-
-  /projects/active:
-    $ref: './openapi-projects.yaml#/paths/~1projects~1active'
+    $ref: './openapi-projects.yaml#/paths/~1projects~1{project_id}~1close'  
 
   /nodes/{nodeInstanceUUID}/outputUi/{outputKey}:
     $ref: './openapi-node-v0.0.1.yaml#/paths/~1nodes~1{nodeInstanceUUID}~1outputUi~1{outputKey}'
