+++ services/web/server/tests/unit/with_dbs/conftest.py
@@ -18,11 +18,13 @@ from pathlib import Path
 import aioredis
 import pytest
 import redis
+import socketio
 import sqlalchemy as sa
 import trafaret_config
 from yarl import URL
 
 import simcore_service_webserver.utils
+from servicelib.rest_responses import unwrap_envelope
 from simcore_service_webserver.application import create_application
 from simcore_service_webserver.application_config import \
     app_schema as app_schema
@@ -197,3 +199,46 @@ async def redis_client(loop, redis_service):
     await client.flushall()
     client.close()
     await client.wait_closed()
+
+
+@pytest.fixture()
+async def socketio_url(client) -> str:
+    SOCKET_IO_PATH = '/socket.io/'
+    return str(client.make_url(SOCKET_IO_PATH))
+
+@pytest.fixture()
+async def security_cookie(client) -> str:
+    # get the cookie by calling the root entrypoint
+    resp = await client.get("/v0/")
+    payload = await resp.json()
+    assert resp.status == 200, str(payload)
+    data, error = unwrap_envelope(payload)
+    assert data
+    assert not error
+
+    cookie = ""
+    if "Cookie" in resp.request_info.headers:
+        cookie = resp.request_info.headers["Cookie"]
+    yield cookie
+
+@pytest.fixture()
+async def socketio_client(socketio_url: str, security_cookie: str):
+    clients = []
+
+    async def connect(client_session_id):
+        sio = socketio.AsyncClient()
+        url = str(URL(socketio_url).with_query({'client_session_id': client_session_id}))
+        await sio.connect(url, headers={'Cookie': security_cookie})
+        assert sio.sid
+        clients.append(sio)
+        return sio
+    yield connect
+    for sio in clients:
+        await sio.disconnect()
+        assert not sio.sid
+
+@pytest.fixture()
+def client_session_id():
+    def create() -> str():
+        return str(uuid4())
+    return create
